views now in project ticket
