--- user_auth/templates/user_auth/index.html
@@ -1,4 +1,0 @@
-{% extends 'user_auth/base.html' %}
-{% block 'content' %}
-	<h2> This is our home page :) its empty </h2>
-{% endblock %}
