--- user_auth/templates/user_auth/profile.html
@@ -1,5 +1,0 @@
-{% extends 'user_auth/base.html' %}
-{% block 'content' %}
-	<h1> Welcome to the profile page </h1>
-	<h3> Hello, {{ user.email }} </h3>
-{% endblock %}
